requirements.txt
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
google-api-core==2.19.1
google-auth==2.31.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
grpcio
grpcio-tools
grpcio-status
idna==3.7
packaging==24.1
proto-plus==1.24.0
protobuf==5.27.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
python-dateutil==2.9.0.post0
requests==2.32.3
rsa==4.9
six==1.16.0
urllib3==2.2.2
mlxtend
upstash-vector==0.5.0
pyyaml
grpcio-reflection
pandas
PyJWT==2.8.0
cryptography==42.0.7
google-cloud-aiplatform
google-generativeai
fuzzywuzzy==0.18.0
db-dtypes==1.3.0
duckdb==1.1.1
Levenshtein==0.26.0
python-Levenshtein==0.26.0